added configuration updater script

+++ root/RucoPie/scripts/start_ui.sh
@@ -1,11 +1,29 @@
 #!/bin/bash
 
-if pgrep -x "love" > /dev/null
-then
-   echo "RucoPie UI is already running, ignoring request."
-else
+source "/root/RucoPie/scripts/commons.sh"
+
+function start () {
+  colorEcho "green" "$separator"
+  colorEcho "green" "Welcome to RucoPie."
+
+  if pgrep -x "love" > /dev/null
+  then
+    colorEcho "green" "RucoPie UI is currently running."
+    return 0
+  fi
+
+  if pgrep -x "retroarch" > /dev/null
+  then
+    colorEcho "green" "Retroarch is currently running."
+    return 0
+  fi
+
   logs="/root/nohup.out"
   [ -e "$logs" ] && rm "$logs"
-  echo "Starting RucoPie UI..."
+  colorEcho "green" "Starting RucoPie UI..."
   nohup love /root/RucoPie/ui > ~/ui.log &
-fi
+
+  colorEcho "green" "$separator"
+}
+
+start
